+++ bundles/org.openhab.voice.habspeaker/src/main/java/org/openhab/voice/habspeaker/internal/io/internal/messages/in/HABSpeakerInInitialize.java
@@ -24,7 +24,7 @@ import org.openhab.voice.habspeaker.internal.config.HABSpeakerThingConfig;
 public class HABSpeakerInInitialize extends HABSpeakerInMessage {
 
     public String id = "";
-    public long sampleRate;
+    public int sampleRate;
 
     public HABSpeakerInInitialize() {
         super(InputCommand.INITIALIZE);
